+++ lib/src/controller/theme_command.dart
@@ -1,7 +1,3 @@
-import 'package:flutter/material.dart';
-
-import 'theme_controller.dart';
-
 /// Reduced API to be exposed of the [ThemeController].
 /// Can use to get the current active theme id,
 /// or to change it using [setTheme(String id)] or [nextTheme()] methods.
@@ -19,12 +15,4 @@ abstract class ThemeCommand {
   /// This returns the [id] parameter of the [AppTheme] instance
   /// that is currently active.
   String get currentThemeId;
-
-  /// Obtains the nearest [ThemeController] up its widget tree and
-  /// returns its value as a [ThemeCommand].
-  ///
-  /// Gets the reference to [ThemeController] but as a reduced version.
-  static ThemeCommand of(BuildContext context) {
-    return ThemeController.of(context);
-  }
 }