+++ lib/src/controller/theme_controller.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../app_theme/app_theme.dart';
+import '../data/app_theme.dart';
 import 'theme_command.dart';
 
 /// Object which controls the behavour of the theme.