+++ lib/theme_provider.dart
@@ -1,9 +1,8 @@
 library theme_provider;
 
-export 'src/app_theme/app_theme_options.dart';
-export 'src/app_theme/app_theme.dart';
-
+export 'src/data/app_theme.dart';
 export 'src/provider/theme_provider.dart';
-export 'src/provider/theme_command.dart';
+
+export 'src/controller/theme_command.dart';
 
 export 'src/widgets/cycle_theme_icon_button.dart';